[ACC-2678] Fold the migrated OpenApiProjectContributorTest into OpenApiSpecBuilderTest, and use application json as input
