+++ contentgrid-appserver-application-model-openapi/build.gradle
@@ -11,4 +11,5 @@ dependencies {
     testImplementation 'org.assertj:assertj-core'
     testImplementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-yaml'
     testImplementation testFixtures(project(':contentgrid-appserver-application-model'))
+    testImplementation project(':contentgrid-appserver-application-model-json')
 }
