+++ contentgrid-appserver-application-model-openapi/src/test/resources/com/contentgrid/appserver/application/model/openapi/specs/scribe-default/openapi.yaml
